Very basic refactor via library and submodule creation
